Move function back to where its used
