+++ apps/back-office/src/app/melden/MeldingForm.tsx
@@ -13,12 +13,11 @@ import type { FormState } from './actions'
 import type { MeldingData } from './types'
 
 import { InvalidFormAlert, LabelsField, PrimaryField, SourceField, SystemErrorAlert, UrgencyField } from './_components'
-import { calculateDefaultValues } from './_utils/calculateDefaultValues'
 import { postMeldingForm } from './actions'
 
 import styles from './MeldingForm.module.css'
 
-export type Props = {
+type Props = {
   defaultValues?: { labels?: number[]; primary?: string; source?: string; urgency?: number }
   existingId?: number
   existingMelding?: MeldingData
@@ -26,6 +25,20 @@ export type Props = {
   labels: LabelOutput[]
   primaryTextArea: StaticFormTextAreaComponentOutput
   sources: SourceOutput[]
+}
+
+// Form components can be prefilled on load on the server, where we fill in existing answers from the backend,
+// or in case of an error, where we use the form data provided.
+// If there is form data, it should take priority over the prefilled components from the server.
+const calculateDefaultValues = (formData?: FormData, defaultValues?: Props['defaultValues']) => {
+  const primaryDefaultValue = (formData?.get('primary') as string | null) ?? defaultValues?.primary ?? ''
+  const sourceDefaultValue = (formData?.get('source') as string | null) ?? defaultValues?.source ?? ''
+  const labelsDefaultValues = formData?.getAll('labels').map((label) => Number(label)) ?? defaultValues?.labels ?? []
+  const rawUrgency = formData?.get('urgency')
+  const urgencyDefaultValue =
+    rawUrgency !== null && rawUrgency !== undefined ? Number(rawUrgency) : (defaultValues?.urgency ?? 0)
+
+  return { labelsDefaultValues, primaryDefaultValue, sourceDefaultValue, urgencyDefaultValue }
 }
 
 const initialState: FormState = {}
